# src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Neo Domotique API",
    "description": "API Backend pour l'application Neo Domotique - Solution de domotique white-label",
    "version": "1.0.0",
    "contact": {
      "name": "Neo Domotique",
      "email": "contact@neo-domotique.fr"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Connexion",
        "description": "Authentifie un utilisateur et retourne les tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "jean.dupont@neo-domotique.fr" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Déconnexion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Déconnexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Déconnexion réussie" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Rafraîchir le token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nouveaux tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": { "type": "string" },
                    "refreshToken": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Utilisateur courant",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Informations utilisateur",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Liste des utilisateurs",
        "description": "Admin only",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "integrateur", "auditeur"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedUsers" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Créer un utilisateur",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUser" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
...
    "/api/projets": {
      "get": {
        "tags": ["Projects"],
        "summary": "Liste des projets",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["brouillon", "en_cours", "termine", "archive"] } },
          { "name": "clientId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedProjects" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Créer un projet",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProject" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        }
      }
    },
...
    "/api/projets/{projectId}/devices": {
      "get": {
        "tags": ["Devices"],
        "summary": "Liste des devices d'un projet",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdPath" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/DeviceWithRoom" }
                }
              }
            }
          }
        }
      }
    },
    "/api/pieces/{roomId}/devices": {
      "get": {
        "tags": ["Devices"],
        "summary": "Liste des devices d'une pièce",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RoomIdPath" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Device" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Devices"],
        "summary": "Créer un device",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RoomIdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDevice" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Device" }
              }
            }
          }
        }
      }
    },
    "/api/devices/{id}": {
      "get": {
        "tags": ["Devices"],
        "summary": "Détail device",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Device" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Devices"],
        "summary": "Modifier device",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateDevice" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Device" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Devices"],
        "summary": "Supprimer device",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Deleted" }
        }
      }
    },
    "/api/pieces/{roomId}/photos": {
      "get": {
        "tags": ["Photos"],
        "summary": "Liste des photos d'une pièce",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RoomIdPath" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Photo" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Photos"],
        "summary": "Upload une photo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RoomIdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "caption": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Photo" }
              }
            }
          }
        }
      }
    },
    "/api/photos/{id}": {
      "delete": {
        "tags": ["Photos"],
        "summary": "Supprimer photo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Deleted" }
        }
      }
    },
...
    "/api/produits": {
      "get": {
        "tags": ["Products"],
        "summary": "Liste des produits",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "brand", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "isActive", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedProducts" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Créer un produit (Admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProduct" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "ProjectIdPath": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "RoomIdPath": {
        "name": "roomId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "default": 1, "minimum": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Non authentifié",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "Deleted": {
        "description": "Supprimé",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Supprimé" }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "code": { "type": "string" }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "integrateur", "auditeur"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateUser": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "integrateur", "auditeur"], "default": "integrateur" }
        }
      },
...
      "PaginatedUsers": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "clientId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["brouillon", "en_cours", "termine", "archive"] },
          "address": { "type": "string" },
          "city": { "type": "string" },
          "postalCode": { "type": "string" },
          "surface": { "type": "number" },
          "roomCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateProject": {
        "type": "object",
        "required": ["clientId", "name"],
        "properties": {
          "clientId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["brouillon", "en_cours", "termine", "archive"], "default": "brouillon" },
          "address": { "type": "string" },
          "city": { "type": "string" },
          "postalCode": { "type": "string" },
          "surface": { "type": "number" },
          "roomCount": { "type": "integer" }
        }
      },
...
      "PaginatedProjects": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        }
      },
...
      "Device": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "roomId": { "type": "string", "format": "uuid" },
          "productId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "serialNumber": { "type": "string" },
          "macAddress": { "type": "string" },
          "ipAddress": { "type": "string" },
          "status": { "type": "string", "enum": ["planifie", "installe", "configure", "operationnel", "en_panne"] },
          "location": { "type": "string" },
          "notes": { "type": "string" },
          "isOnline": { "type": "boolean" },
          "lastSeenAt": { "type": "string", "format": "date-time" },
          "installedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "DeviceWithRoom": {
        "allOf": [
          { "$ref": "#/components/schemas/Device" },
          {
            "type": "object",
            "properties": {
              "roomName": { "type": "string" },
              "roomType": { "type": "string" }
            }
          }
        ]
      },
      "CreateDevice": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "productId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "serialNumber": { "type": "string" },
          "macAddress": { "type": "string" },
          "ipAddress": { "type": "string" },
          "status": { "type": "string", "enum": ["planifie", "installe", "configure", "operationnel", "en_panne"], "default": "planifie" },
          "location": { "type": "string" },
          "notes": { "type": "string" }
        }
      },
      "UpdateDevice": {
        "type": "object",
        "properties": {
          "productId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "serialNumber": { "type": "string" },
          "macAddress": { "type": "string" },
          "ipAddress": { "type": "string" },
          "status": { "type": "string", "enum": ["planifie", "installe", "configure", "operationnel", "en_panne"] },
          "location": { "type": "string" },
          "notes": { "type": "string" },
          "isOnline": { "type": "boolean" }
        }
      },
      "Photo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "roomId": { "type": "string", "format": "uuid" },
          "filename": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "caption": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "reference": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "brand": { "type": "string" },
          "priceHT": { "type": "number" },
          "tvaRate": { "type": "number" },
          "imageUrl": { "type": "string", "format": "uri" },
          "isActive": { "type": "boolean" },
          "stock": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateProduct": {
        "type": "object",
        "required": ["reference", "name", "category", "priceHT"],
        "properties": {
          "reference": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "brand": { "type": "string" },
          "priceHT": { "type": "number" },
          "tvaRate": { "type": "number", "default": 20 },
          "imageUrl": { "type": "string", "format": "uri" },
          "isActive": { "type": "boolean", "default": true },
          "stock": { "type": "integer" }
        }
      },
...
      "PaginatedProducts": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        }
      },
...
    }
  }
}